requirements.txt
streamlit
google-generativeai
python-dotenv
Pillow
PyMuPDF
plotly
